fix(cosmic-swingset): finish reorganization
