+++ packages/cosmic-swingset/jsconfig.json
@@ -14,5 +14,5 @@
     "strictNullChecks": true,
     "moduleResolution": "node",
   },
-  "include": ["lib/**/*.js", "exported.js", "globals.d.ts"],
+  "include": ["src/**/*.js", "exported.js"],
 }
