+++ packages/cosmic-swingset/package.json
@@ -5,7 +5,10 @@
   "parsers": {
     "js": "mjs"
   },
-  "main": "lib/ag-solo/main.js",
+  "bin": {
+    "ag-chain-cosmos": "./src/chain-entrypoint.cjs"
+  },
+  "main": "src/chain-main.js",
   "repository": "https://github.com/Agoric/agoric-sdk",
   "scripts": {
     "build": "exit 0",
@@ -25,53 +28,20 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@agoric/assert": "^0.2.9",
-    "@agoric/babel-parser": "^7.6.4",
-    "@agoric/bundle-source": "^1.3.4",
-    "@agoric/captp": "^1.7.10",
     "@agoric/cosmos": "^0.25.4",
     "@agoric/dapp-svelte-wallet": "^0.9.4",
-    "@agoric/ertp": "^0.10.8",
-    "@agoric/eventual-send": "^0.13.11",
     "@agoric/import-bundle": "^0.2.11",
     "@agoric/install-metering-and-ses": "^0.2.10",
-    "@agoric/install-ses": "^0.5.10",
-    "@agoric/marshal": "^0.4.8",
-    "@agoric/nat": "^4.0.0",
-    "@agoric/notifier": "^0.3.11",
-    "@agoric/pegasus": "^0.2.4",
-    "@agoric/promise-kit": "^0.2.10",
-    "@agoric/registrar": "^0.2.11",
-    "@agoric/same-structure": "^0.1.10",
-    "@agoric/sharing-service": "^0.1.11",
-    "@agoric/sparse-ints": "^0.1.6",
-    "@agoric/spawner": "^0.4.11",
     "@agoric/store": "^0.4.11",
     "@agoric/swing-store-lmdb": "^0.4.9",
-    "@agoric/swing-store-simple": "^0.3.6",
     "@agoric/swingset-vat": "^0.16.4",
-    "@agoric/transform-eventual-send": "^1.4.6",
-    "@agoric/treasury": "^0.3.4",
-    "@agoric/zoe": "^0.15.4",
-    "@babel/generator": "^7.6.4",
+    "@agoric/vats": "^0.1.0",
     "@iarna/toml": "^2.2.3",
     "@opentelemetry/exporter-prometheus": "^0.16.0",
     "@opentelemetry/metrics": "^0.16.0",
     "agoric": "^0.12.13",
     "anylogger": "^0.21.0",
-    "chalk": "^2.4.2",
-    "deterministic-json": "^1.0.5",
-    "esm": "^3.2.25",
-    "express": "^4.17.1",
-    "minimist": "^1.2.0",
-    "morgan": "^1.9.1",
-    "n-readlines": "^1.0.0",
-    "node-fetch": "^2.6.0",
-    "node-lmdb": "^0.9.4",
-    "polycrc": "https://github.com/agoric-labs/node-polycrc",
-    "rollup": "^1.26.2",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "temp": "^0.9.1",
-    "ws": "^7.2.0"
+    "esm": "^3.2.25"
   },
   "devDependencies": {
     "ava": "^3.12.1",
